refactor(state): split state.rs into per-domain submodules
The 23 types previously colocated in src/state.rs make "edit one
impl, miss another" likely when an AI agent searches a 797-line
file with multiple impls per type. Split into 12 per-domain
submodules under src/state/ averaging ~60 lines each (largest 146).

state/mod.rs re-exports every type via pub(crate) use, so all 158
existing crate::state::Foo call sites compile unchanged.
ease_out_quad drops from pub(crate) to private — only roulette.rs
calls it.

No behavioral change. cargo build / clippy --all-targets -D warnings
/ nightly fmt --check / test --workspace all clean (1021 tests +
1 doctest pass).
